flexmind-ros1-release-1.0.0/src/fleximind_remote/src/fleximind_remote/requirements.txt
# keep in alphabetical order
dm_control
pyrealsense2
quaternion
tyro
numpy-quaternion
